+++ src/Models/HubbardOneBand/ModelHubbard.h
@@ -110,7 +110,7 @@ public:
 	typedef TargetQuantumElectrons<RealType> TargetQuantumElectronsType;
 	typedef typename ModelBaseType::SparseMatrixType SparseMatrixType;
 	typedef typename ModelHelperType::SparseElementType SparseElementType;
-	typedef unsigned int long long WordType;
+	typedef unsigned int long WordType;
 	typedef  HilbertSpaceHubbard<WordType> HilbertSpaceHubbardType;
 	typedef typename ModelBaseType::VectorOperatorType VectorOperatorType;
 	typedef typename ModelBaseType::SymmetryElectronsSzType SymmetryElectronsSzType;
